backend/requirements.txt
flask
flask-cors
pandas
numpy
requests
firebase-admin
python-dotenv
